requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# api/api.py: 1
Flask == 1.0.2

# core/sast/java.py: 6
beautifulsoup4 == 4.7.1

# core/vcs/bitbucket.py: 2
bitbucket_python == 0.2.2

# core/sast/java.py: 5
# file.py: 2
# main.py: 9
colorama == 0.3.9

# core/utils/old/defectdojo.py: 1
defectdojo_api == 1.1.4

# core/utils/elastic.py: 2
elasticsearch == 7.0.2

# core/utils/old/jsonmerge.py: 2
flatten_json == 0.1.7

# core/utils/json_parsing.py: 3,4,5
# core/utils/old/slack_webhook.py: 2,3,4
# core/utils/test.py: 3,4,5
# core/utils/utils.py: 2,3,5
mysql_connector_python == 8.0.17

# core/vcs/bitbucket.py: 10
pygit2 == 0.28.0

# core/utils/json_parsing.py: 7
# core/utils/old/slack_webhook.py: 7
# core/utils/test.py: 7
# main.py: 15
requests == 2.21.0

# core/sast/java.py: 9
xmltodict == 0.12.0
